requirements.txt
Flask>=3.0,<4.0
Flask-Login>=0.6,<1.0
Flask-SQLAlchemy>=3.1,<4.0
gunicorn>=21,<23
psycopg2-binary>=2.9,<3.0
